LPS-82018 delete friendlyUrls only if they exist

+++ modules/apps/journal/journal-service/src/main/java/com/liferay/journal/service/impl/JournalArticleLocalServiceImpl.java
@@ -1242,9 +1242,19 @@ public class JournalArticleLocalServiceImpl
 
 		// FriendlyURL
 
-		friendlyURLEntryLocalService.deleteFriendlyURLEntry(
-			article.getGroupId(), JournalArticle.class,
-			article.getResourcePrimKey());
+		long classNameId = classNameLocalService.getClassNameId(
+			JournalArticle.class);
+
+		List<FriendlyURLEntry> friendlyURLEntries =
+			friendlyURLEntryLocalService.getFriendlyURLEntries(
+				article.getGroupId(), classNameId,
+				article.getResourcePrimKey());
+
+		if (!friendlyURLEntries.isEmpty()) {
+			friendlyURLEntryLocalService.deleteFriendlyURLEntry(
+				article.getGroupId(), JournalArticle.class,
+				article.getResourcePrimKey());
+		}
 
 		// Article
